+++ apps/frontend/src/modules/courses/slices/courses.slice.ts
@@ -4,14 +4,16 @@ import { DataStatus } from "~/libs/enums/enums.js";
 import { type ValueOf } from "~/libs/types/types.js";
 
 import { type CourseDto } from "../libs/types/types.js";
-import { getAll } from "./actions.js";
+import { getAll, getRecommended } from "./actions.js";
 
 type State = {
+	recommendedCourses: CourseDto[];
 	searchDataStatus: ValueOf<typeof DataStatus>;
 	searchedCourses: CourseDto[];
 };
 
 const initialState: State = {
+	recommendedCourses: [],
 	searchDataStatus: DataStatus.IDLE,
 	searchedCourses: [],
 };
@@ -26,6 +28,16 @@ const { actions, name, reducer } = createSlice({
 			state.searchDataStatus = DataStatus.PENDING;
 		});
 		builder.addCase(getAll.rejected, (state) => {
+			state.searchDataStatus = DataStatus.REJECTED;
+		});
+		builder.addCase(getRecommended.fulfilled, (state, action) => {
+			state.recommendedCourses = action.payload.courses;
+			state.searchDataStatus = DataStatus.FULFILLED;
+		});
+		builder.addCase(getRecommended.pending, (state) => {
+			state.searchDataStatus = DataStatus.PENDING;
+		});
+		builder.addCase(getRecommended.rejected, (state) => {
 			state.searchDataStatus = DataStatus.REJECTED;
 		});
 	},